Refactoring of CTracker class
